# [perf] User page 성능 최적화 - TBT 51% 감소, CLS 70% 개선