… .eslintrc.json
… PERFORMANCE_OPTIMIZATION.md
… Portfolio_Onboarding_Optimization.md
… Portfolio_UserPage_Optimization.md
… UserPage_Optimization_Results.md
… lighthouse-after.json
… lighthouse-after2.json
… lighthouse-analysis.json
… lighthouse-detail.json
… lighthouse-home.json
… lighthouse-lcp-debug.json
… lighthouse-real.json
… lighthouse-report.html
… lighthouse-report.json
… lighthouse-test.json
… lighthouse-user-page.json
… package.json
… scripts/simple-perf-test.js
… src/app/page.tsx
… src/app/users/[userId]/page-backup.tsx
… src/components/features/auth/AuthModal.tsx
… src/components/features/auth/forms/SignUpForm.tsx
… src/components/features/auth/index.ts
… src/components/features/career/CareerControlPanel.tsx
… src/components/features/career/CareerDashboard.tsx
… src/components/features/career/CareerTimeline/components/EmptyState.tsx
… src/components/features/career/CareerTimeline/index.tsx
… src/components/features/career/LazyCareerDashboard.tsx
… src/components/features/dashboard/AiCareerInsights.tsx
… src/components/features/dashboard/CompensationAnalysis/components/AIInsights.tsx