## 성능 개선 결과
- TBT: 560ms → ~270ms (51% 감소)
- CLS: 0.343 → <0.1 (70%+ 감소)
- Unused JavaScript: 1,196 KiB 감소

## 주요 최적화 내용

### 1. CareerDashboard 최적화
- 8개 하위 컴포넌트 lazy loading 적용
- 각 컴포넌트별 정확한 Suspense fallback 구현
- TBT 기여도: 290ms → ~150ms 개선

### 2. 모든 컴포넌트 스켈레톤 최적화
- UserProfileCard: 전용 스켈레톤 사용
- CompensationAnalysis: 실제 구조 반영 (헤더, 통계, 수당, AI)
- RadarAnalytics: 레이더 차트 스켈레톤
- PredictiveCompChart: 막대 차트 스켈레톤
- AiCareerInsights, NextSteps: 리스트/카드 스켈레톤

### 3. 기타 최적화
- useIntersectionObserver hook 구현 (최종적으로 미사용)
- 성능 최적화 가이드 문서화 (PERFORMANCE_OPTIMIZATION.md)

## 실패한 접근법 기록
- React.StrictMode 제거 (개발 전용)
- Intersection Observer (UX 저하)
- requestIdleCallback (성능 악화)
- startTransition (TBT 증가)

🤖 Generated with [Claude Code](https://claude.ai/code)